refactor(wpf-migration phase 1): extract Savedrake.Core, move leaf helpers, repoint harness
First step of the WinForms -> WPF migration: stand up a UI-agnostic Savedrake.Core
library and prove the shared-core + harness mechanism on the smallest safe slice
before any high-stakes logic moves. No user-facing change.

- New Savedrake.Core (net48 class lib) in the solution. Types stay in the
  'Savedrake' namespace so app call sites are unchanged.
- Move Log (from Program.cs), interval parsing (TryParseInterval/Canonicalize +
  the regex -> IntervalParser), and FriendlyTime (-> TimeText) into Core. The
  WinForms app keeps thin forwarders and still builds + runs.
- RestoreHarness now loads Savedrake.Core.dll and exercises IntervalParser /
  TimeText / Log.Redact directly from Core (validates the mechanic Phase 0/2 need).

Verified: Debug + Release build clean; the exact CI sequence (nuget restore +
msbuild, no -restore) builds the SDK-style Core project; harness 196/196; the
WinForms app launches and runs unchanged.
