+++ Savedrake.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Savedrake-Updater", "update
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RestoreHarness", "RestoreHarness\RestoreHarness.csproj", "{097EDA73-C89A-4DC7-A5F2-0A79CBD658FD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Savedrake.Core", "Savedrake.Core\Savedrake.Core.csproj", "{B2E4B7A1-3C5D-4E6F-A1B2-C3D4E5F60718}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{097EDA73-C89A-4DC7-A5F2-0A79CBD658FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{097EDA73-C89A-4DC7-A5F2-0A79CBD658FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{097EDA73-C89A-4DC7-A5F2-0A79CBD658FD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B2E4B7A1-3C5D-4E6F-A1B2-C3D4E5F60718}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B2E4B7A1-3C5D-4E6F-A1B2-C3D4E5F60718}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B2E4B7A1-3C5D-4E6F-A1B2-C3D4E5F60718}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B2E4B7A1-3C5D-4E6F-A1B2-C3D4E5F60718}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
